Refactoring Inference Management UI (#193380)
## Summary

This PR includes:
- Adding header links and text
- Restructure of columns
- Increased pagination limit

![Screenshot 2024-09-18 at 11 51
25 PM](https://github.com/user-attachments/assets/a4fc1324-4365-4f1c-a33e-3d247227bcfa)

### Checklist

Delete any items that are not applicable to this PR.

- [X] Any text added follows [EUI's writing
guidelines](https://elastic.github.io/eui/#/guidelines/writing), uses
sentence case text and includes [i18n
support](https://github.com/elastic/kibana/blob/main/packages/kbn-i18n/README.md)
- [X] [Unit or functional
tests](https://www.elastic.co/guide/en/kibana/master/development-tests.html)
were updated or added to match the most common scenarios

---------

(cherry picked from commit fd149a5ab2c935d9f8f2dfe17c7f35c8a1c1f353)
