+++ x-pack/plugins/search_inference_endpoints/public/components/all_inference_endpoints/constants.ts
@@ -14,7 +14,7 @@ import {
 } from './types';
 
 export const DEFAULT_TABLE_ACTIVE_PAGE = 1;
-export const DEFAULT_TABLE_LIMIT = 10;
+export const DEFAULT_TABLE_LIMIT = 25;
 
 export const DEFAULT_QUERY_PARAMS: QueryParams = {
   page: DEFAULT_TABLE_ACTIVE_PAGE,
