+++ x-pack/plugins/search_inference_endpoints/public/components/all_inference_endpoints/render_table_columns/render_service_provider/translations.ts
@@ -7,18 +7,6 @@
 
 import { i18n } from '@kbn/i18n';
 
-export const THREADS = (numThreads: number) =>
-  i18n.translate('xpack.searchInferenceEndpoints.elasticsearch.threads', {
-    defaultMessage: 'Threads: {numThreads}',
-    values: { numThreads },
-  });
-
-export const ALLOCATIONS = (numAllocations: number) =>
-  i18n.translate('xpack.searchInferenceEndpoints.elasticsearch.allocations', {
-    defaultMessage: 'Allocations: {numAllocations}',
-    values: { numAllocations },
-  });
-
 export const MIT_LICENSE = i18n.translate(
   'xpack.searchInferenceEndpoints.elasticsearch.mitLicense',
   {
