+++ x-pack/plugins/search_inference_endpoints/public/components/inference_endpoints_header.tsx
@@ -5,7 +5,7 @@
  * 2.0.
  */
 
-import { EuiPageTemplate, EuiLink, EuiText, EuiSpacer } from '@elastic/eui';
+import { EuiPageTemplate, EuiLink } from '@elastic/eui';
 import React from 'react';
 import * as i18n from '../../common/translations';
 import { docLinks } from '../../common/doc_links';
@@ -18,21 +18,16 @@ export const InferenceEndpointsHeader: React.FC = () => {
     <EuiPageTemplate.Header
       data-test-subj="allInferenceEndpointsPage"
       pageTitle={i18n.INFERENCE_ENDPOINT_LABEL}
-      description={
-        <EuiText>
-          {i18n.MANAGE_INFERENCE_ENDPOINTS_LABEL}
-          <EuiSpacer size="s" />
-          <EuiLink
-            href={docLinks.createInferenceEndpoint}
-            target="_blank"
-            data-test-subj="learn-how-to-create-inference-endpoints"
-          >
-            {i18n.LEARN_HOW_TO_CREATE_INFERENCE_ENDPOINTS_LINK}
-          </EuiLink>
-        </EuiText>
-      }
+      description={i18n.MANAGE_INFERENCE_ENDPOINTS_LABEL}
       bottomBorder={true}
       rightSideItems={[
+        <EuiLink
+          href={docLinks.createInferenceEndpoint}
+          target="_blank"
+          data-test-subj="api-documentation"
+        >
+          {i18n.API_DOCUMENTATION_LINK}
+        </EuiLink>,
         <EuiLink href={trainedModelPageUrl} target="_blank" data-test-subj="view-your-models">
           {i18n.VIEW_YOUR_MODELS_LINK}
         </EuiLink>,
