+++ x-pack/plugins/search_inference_endpoints/public/hooks/use_table_data.test.tsx
@@ -118,9 +118,7 @@ describe('useTableData', () => {
       b.inference_id.localeCompare(a.inference_id)
     );
 
-    const sortedEndpoints = result.current.sortedTableData.map(
-      (item) => item.endpoint.inference_id
-    );
+    const sortedEndpoints = result.current.sortedTableData.map((item) => item.endpoint);
     const expectedModelIds = expectedSortedData.map((item) => item.inference_id);
 
     expect(sortedEndpoints).toEqual(expectedModelIds);
@@ -153,19 +151,6 @@ describe('useTableData', () => {
       { wrapper }
     );
     const filteredData = result.current.sortedTableData;
-    expect(
-      filteredData.every((item) => item.endpoint.inference_id.includes(searchKey))
-    ).toBeTruthy();
-  });
-
-  it('should update deployment status based on deploymentStatus object', () => {
-    const { result } = renderHook(
-      () => useTableData(inferenceEndpoints, queryParams, filterOptions, searchKey),
-      { wrapper }
-    );
-
-    const updatedData = result.current.sortedTableData;
-
-    expect(updatedData[2].deployment).toEqual('started');
+    expect(filteredData.every((item) => item.endpoint.includes(searchKey))).toBeTruthy();
   });
 });
