+++ x-pack/plugins/search_inference_endpoints/public/hooks/use_table_data.tsx
@@ -7,7 +7,7 @@
 
 import type { EuiTableSortingType } from '@elastic/eui';
 import { Pagination } from '@elastic/eui';
-import { DeploymentState, InferenceAPIConfigResponse } from '@kbn/ml-trained-models-utils';
+import { InferenceAPIConfigResponse } from '@kbn/ml-trained-models-utils';
 import { useMemo } from 'react';
 import { TaskTypes } from '../../common/types';
 import { DEFAULT_TABLE_LIMIT } from '../components/all_inference_endpoints/constants';
@@ -19,7 +19,6 @@ import {
   SortOrder,
   ServiceProviderKeys,
 } from '../components/all_inference_endpoints/types';
-import { useTrainedModelStats } from './use_trained_model_stats';
 
 interface UseTableDataReturn {
   tableData: InferenceEndpointUI[];
@@ -35,15 +34,6 @@ export const useTableData = (
   filterOptions: FilterOptions,
   searchKey: string
 ): UseTableDataReturn => {
-  const { data: trainedModelStats } = useTrainedModelStats();
-
-  const deploymentStatus = trainedModelStats?.trained_model_stats.reduce((acc, modelStat) => {
-    if (modelStat.deployment_stats?.deployment_id) {
-      acc[modelStat.deployment_stats.deployment_id] = modelStat?.deployment_stats?.state;
-    }
-    return acc;
-  }, {} as Record<string, DeploymentState | undefined>);
-
   const tableData: InferenceEndpointUI[] = useMemo(() => {
     let filteredEndpoints = inferenceEndpoints;
 
@@ -61,21 +51,12 @@ export const useTableData = (
 
     return filteredEndpoints
       .filter((endpoint) => endpoint.inference_id.includes(searchKey))
-      .map((endpoint) => {
-        const isElasticService =
-          endpoint.service === ServiceProviderKeys.elasticsearch ||
-          endpoint.service === ServiceProviderKeys.elser;
-        const deploymentId = isElasticService ? endpoint.inference_id : undefined;
-        const deployment = (deploymentId && deploymentStatus?.[deploymentId]) || undefined;
-
-        return {
-          deployment,
-          endpoint,
-          provider: endpoint.service,
-          type: endpoint.task_type,
-        };
-      });
-  }, [inferenceEndpoints, searchKey, filterOptions, deploymentStatus]);
+      .map((endpoint) => ({
+        endpoint: endpoint.inference_id,
+        provider: endpoint,
+        type: endpoint.task_type,
+      }));
+  }, [inferenceEndpoints, searchKey, filterOptions]);
 
   const sortedTableData: InferenceEndpointUI[] = useMemo(() => {
     return [...tableData].sort((a, b) => {
@@ -83,9 +64,9 @@ export const useTableData = (
       const bValue = b[queryParams.sortField];
 
       if (queryParams.sortOrder === SortOrder.asc) {
-        return aValue.inference_id.localeCompare(bValue.inference_id);
+        return aValue.localeCompare(bValue);
       } else {
-        return bValue.inference_id.localeCompare(aValue.inference_id);
+        return bValue.localeCompare(aValue);
       }
     });
   }, [tableData, queryParams]);
